Adding CKEditor to the Post Form
I wanted to give more control of the post style to the user, so I used the ckeditor package provided by a third
party owner in Django, I used the command pip install django-ckeditor. Then, I had to put two apps in INSTALLED_APS
at settings.py, and also the CKEDITOR_UPLOAD_PATH to indicate where ckeditor is going to save the post content
images. In urls.py I added a new URLconf provided by ckeditor in order to handle that logic of uploading an image.

The, I had to create the form in forms.py, specifying the widget as CKEditorUploadingWidget() to style the initial
textarea with the ckeditor widget. I added a clean method that checks if the main post image size is greater than
the size allowed. Finally, in views.py I added the form_class = PostForm attribute to indicate the class which form
to use.

In views.py from the users app, I added one line to make all the new users with 'is_staff' attribute as true, given
that ckeditor only allows staff users to upload images. There was a bug that deleted the post image when the user
is updating the post and doesn't include a new image. In orfer to fix this, I checked the previous and new instances,
if they have the same image path, it means the user didn't upload a new image so my server doesn't need to
remove the previous image.

+++ django_project/blog/signals.py
@@ -14,9 +14,9 @@ def file_update(sender, instance, raw, using, update_fields,**kwargs):
         pass
 
     else:
-        # current = instance
         previous = Post.objects.get(id=instance.id)
-        os.remove(previous.image.path)
+        if instance.image.path != previous.image.path:
+            os.remove(previous.image.path)
 
 
 @receiver(post_save, sender=Comment)
